ajout de route avec requêtes dans mongo selon role

+++ src/routes/index.js
@@ -1,27 +1,22 @@
 const express = require('express')
 const _ = require('lodash')
-const jwt = require('jsonwebtoken')
 const bcrypt = require('bcryptjs')
 const db = require('../mongo')
-const { jwtSecretKey } = require('../config')
 
-// on créer une instance d'express router
+// on crée une instance d'express router
 const router = express.Router()
-
-// génération d'un token
-function generateToken(username) {
-    jwt.sign({ username }, jwtSecretKey)
-}
 
 // récupération du token
 function getToken(req) {
     _.replace(req.headers.authorization, 'Bearer ', '')
 }
 
+/**
+ * WELOME ROUTE
+ */
 // router pour vérfier si le back marche
 router.get('/welcome', async (req, res) => {
     res.json({ message: 'Welcome to the API!' })
-
     console.log(`OK ça marche`)
 })
 
@@ -41,7 +36,7 @@ router.get('/me', async (req, res) => {
     res.json(user)
 })
 
-// on dit a nos routes d'utiliser a travers un préfix une autre instance d'express router.
+// on crée une route pour se connecter
 router.post('/login', async (req, res) => {
     const { username, password } = req.body // on récupère le username et le password dans le body de la requête
 
@@ -160,44 +155,69 @@ router.post('/create-assignment', async (req, res) => {
     console.log(`A user created assignment ${nom}.`)
 
     res.json(assignment) // on renvoie l'assignment au client
+    next() // on passe à la route suivante
 })
 
 // route pour modifier un assignment
 router.put('/update-assignment/:id', async (req, res) => {
-    const { id } = req.params // on récupère l'id dans les params de la requête
+    const token = getToken(req) // on récupère le token dans le header de la requête
+    const role = await db.users.findRoleByToken(token) // on récupère les informations de l'utilisateur
 
-    const { nom, auteur, matiere, dateDeRendu, note, rendu, remarque } =
-        req.body // on récupère les données à modifier dans le body de la requête
+    if (role === 'TEACHER' || role === 'ADMIN') {
+        // si l'utilisateur est un prof
+        const { id } = req.params // on récupère l'id dans les params de la requête
 
-    const assignment = await db.assignments.updateAssignmentById(
-        // on modifie l'assignment
-        {
-            nom,
-            auteur,
-            matiere,
-            dateDeRendu,
-            note,
-            rendu,
-            remarque,
-        },
-        id
-    )
+        const { nom, auteur, matiere, dateDeRendu, note, rendu, remarque } =
+            req.body // on récupère les données à modifier dans le body de la requête
 
-    console.log(`A user updated assignment ${nom}.`)
+        const assignment = await db.assignments.updateAssignmentById(
+            // on modifie l'assignment
+            {
+                nom,
+                auteur,
+                matiere,
+                dateDeRendu,
+                note,
+                rendu,
+                remarque,
+            },
+            id
+        )
 
-    res.json(assignment) // on renvoie l'assignment au client
+        console.log(`A user updated assignment ${nom}.`)
+
+        res.json(assignment) // on renvoie l'assignment au client
+        next() // on passe à la route suivante
+    } else {
+        // si l'utilisateur n'est pas un prof
+        const errorMessage = `User is not authorized to perform this action.`
+        console.log(errorMessage)
+        return res.status(401).json({ message: errorMessage })
+    }
 })
 
 // route pour supprimer un assignment
 router.delete('/delete-assignment/:id', async (req, res) => {
-    const { id } = req.params // on récupère l'id dans les params de la requête
+    const token = getToken(req) // on récupère le token dans le header de la requête
+    const role = await db.users.findRoleByToken(token) // on récupère les informations de l'utilisateur
 
-    const assignment = await db.assignments.deleteAssignment(id) // on supprime l'assignment
+    if (role === 'TEACHER' || role === 'ADMIN') {
+        // si l'utilisateur est un prof
 
-    console.log(`A user deleted assignment ${assignment.nom}.`)
-    res.json(assignment) // on renvoie l'assignment au client
+        const { id } = req.params // on récupère l'id dans les params de la requête
+
+        const assignment = await db.assignments.deleteAssignment(id) // on supprime l'assignment
+
+        console.log(`A user deleted assignment ${assignment.nom}.`)
+        res.json(assignment) // on renvoie l'assignment au client
+        next() // on passe à la route suivante
+    } else {
+        // si l'utilisateur n'est pas un prof
+        const errorMessage = `User is not authorized to perform this action.`
+        console.log(errorMessage)
+        return res.status(401).json({ message: errorMessage })
+    }
 })
-
 
 
 module.exports = router
